Add Refactor Light and Orientation to use BaseRecord
