+++ android/src/main/java/com/sensormanager/ManagerSensorModule.kt
@@ -1,30 +1,37 @@
 package com.sensormanager
 
-import android.os.Build
-import androidx.annotation.RequiresApi
-import com.facebook.react.bridge.*
+import com.facebook.react.bridge.ReactApplicationContext
+import com.facebook.react.bridge.ReactContextBaseJavaModule
+import com.facebook.react.bridge.ReactMethod
 
 
 class SensorManagerModule(reactContext: ReactApplicationContext) : ReactContextBaseJavaModule(reactContext) {
   private val orientationRecord: OrientationRecord = OrientationRecord(reactContext)
+  private val lightRecord: LightRecord = LightRecord(reactContext)
+  private val defaultRefresh: Int = 1000
 
   override fun getName(): String {
         return "SensorManager"
     }
 
-    @RequiresApi(Build.VERSION_CODES.KITKAT)
     @ReactMethod
     fun startOrientation(freq: Int?) {
-      if (freq !== null) {
-        orientationRecord.startTrack(freq)
-      } else {
-        orientationRecord.startTrack(1000)
-      }
+      orientationRecord.start(freq ?: defaultRefresh)
     }
 
     @ReactMethod
     fun stopOrientation() {
-      orientationRecord.stopTrack()
+      orientationRecord.stop()
+    }
+
+    @ReactMethod
+    fun startLight(freq: Int?) {
+      lightRecord.start(freq ?: defaultRefresh)
+    }
+
+    @ReactMethod
+    fun stopLight() {
+      lightRecord.stop()
     }
 
 }
